:chart_with_upwards_trend: GENERAL - fixed analytics
