+++ utils/gtag.js
@@ -1,31 +1,26 @@
 import { omit, set, isFunction, chunk, forEach, get, clone } from "lodash";
 
 const TagManager = {
-  init: function (googleAnalyticsCode = "") {
-    if (!googleAnalyticsCode) {
+  init: function (googleTagManagerCode = "") {
+    if (!googleTagManagerCode) {
       return null;
     }
 
     return (
       <>
-        {/* Global Site Tag (gtag.js) - Google Analytics */}
-        <script
-          async
-          src={`https://www.googletagmanager.com/gtag/js?id=${googleAnalyticsCode}`}
-        />
+        {/* Google Tag Manager */}
         <script
           dangerouslySetInnerHTML={{
             __html: `
-                window.dataLayer = window.dataLayer || [];
-                window.gtmTrackingId = '${googleAnalyticsCode}';
-                function gtag(){dataLayer.push(arguments);}
-                gtag('js', new Date());
-                gtag('config', '${googleAnalyticsCode}', {
-                    page_path: window.location.pathname,
-                });
+              (function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
+              new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
+              j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
+              'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
+              })(window,document,'script','dataLayer','${googleTagManagerCode}');
             `,
           }}
         />
+        <noscript><iframe src={`https://www.googletagmanager.com/ns.html?id=${googleTagManagerCode}`} height="0" width="0" style={{ display:'none', visibility:'hidden'}}></iframe></noscript>
       </>
     );
   },
